+++ test/test_common_libs.dart
@@ -1,0 +1,2 @@
+export 'package:flutter_test/flutter_test.dart';
+export 'package:boolean_validation/boolean_validation.dart';