+++ CHANGELOG.md
@@ -28,3 +28,15 @@
 
 - Provided an example Flutter app demonstrating how to use the validation methods in a form.
 
+## [0.0.2] - 2024-09-24
+
+- support for version 3.3.2 in flutter
+
+## [0.1.0] - 2024-09-29
+
+1- Add constrained email validation for specific providers like gmail, yahoo, etc.
+2- add more advanced password validation for complex stages if developers required them
+3- add lat and long validations
+4- add validation for Full Name, First and Last Name
+5- add Validations for required fields
+6- add alpha and alpha numeric validations